Revert CSP implementation
